chore(profile): Move profile display to profile page
