+++ packages/app/components/Card/Card.tsx
@@ -2,7 +2,7 @@ import { styled } from "@/stitches.config";
 
 const Card = styled("div", {
   margin: "0 auto",
-  marginBottom: "$12",
+  marginBottom: 64,
   maxWidth: 320,
 });
 export default Card;
