+++ packages/app/components/Post/PostList.tsx
@@ -1,0 +1,7 @@
+import { styled } from "@/stitches.config";
+
+export const PostList = styled("div", {
+  display: "flex",
+  flexDirection: "column",
+  rowGap: "$4",
+});
