+++ packages/app/components/Post/index.ts
@@ -1,2 +1,3 @@
+export * from "./PostList";
 export * from "./Post";
 export { default } from "./Post";
