+++ packages/app/components/TabBar/index.tsx
@@ -1,0 +1,2 @@
+export * from "./TabBar";
+export { default } from "./TabBar";
