+++ packages/app/components/Title/Title.tsx
@@ -1,0 +1,7 @@
+import { styled } from "@/stitches.config";
+
+const Title = styled("h2", {
+  marginTop: "$12",
+  marginBottom: "$5",
+});
+export default Title;
