+++ packages/app/components/Title/index.ts
@@ -1,0 +1,2 @@
+export * from "./Title";
+export { default } from "./Title";
