+++ packages/app/pages/index.tsx
@@ -3,57 +3,11 @@ import { styled } from "@/stitches.config";
 import { ChevronDownIcon } from "@radix-ui/react-icons";
 
 import { Button, Heading, Input } from "@cabindao/topo";
-import Card from "@/components/Card";
-import UserCard from "@/components/UserCard";
-import Post from "@/components/Post";
+import Post, { PostList } from "@/components/Post";
 import { useWallet } from "@/components/WalletAuth";
+import Title from "@/components/Title";
+import TabBar, { TabButton } from "@/components/TabBar";
 
-const Title = styled("h2", {
-  marginTop: "$12",
-  marginBottom: "$5",
-});
-
-const TabBarWrapper = styled("div", {
-  borderBottomWidth: 1,
-  borderBottomStyle: "solid",
-  borderColor: "$forest",
-  boxSizing: "border-box",
-  marginBottom: "$2",
-});
-const TabBarContent = styled("div", {
-  display: "flex",
-  marginBottom: -1,
-});
-const TabLink = styled("button", {
-  boxSizing: "border-box",
-  background: "none",
-  paddingLeft: "$4",
-  paddingRight: "$4",
-  border: 0,
-  cursor: "pointer",
-  defaultVariants: {
-    active: false,
-  },
-  variants: {
-    active: {
-      false: {
-        "&:hover": {
-          backgroundColor: "rgba(50, 72, 65, 0.1)",
-        },
-      },
-      true: {
-        backgroundColor: "$forest",
-        color: "$sand",
-      },
-    },
-  },
-});
-const TabButton = (props: any) => <TabLink {...props} />;
-const TabBar = ({ children, ...props }: { children?: React.ReactNode }) => (
-  <TabBarWrapper {...props}>
-    <TabBarContent>{children}</TabBarContent>
-  </TabBarWrapper>
-);
 const MenuButton = (props: any) => {
   return (
     <Button
@@ -69,31 +23,15 @@ const MenuButton = (props: any) => {
   );
 };
 
-const PostList = styled("div", {
-  display: "flex",
-  flexDirection: "column",
-  rowGap: "$4",
-});
-
 const Home: NextPage = () => {
   const { address, ens } = useWallet();
   return (
     <div>
       <header>
-        {address && (
-          <>
-            <Title>Profile</Title>
-            <Card>
-              <UserCard address={address} ens={ens} />
-            </Card>
-          </>
-        )}
-
         <Title>Today</Title>
 
         <TabBar>
-          <TabButton active>Submissions</TabButton>
-          <TabButton>Upvoted</TabButton>
+          <TabButton active>Links</TabButton>
           <div style={{ marginLeft: "auto" }}>
             <MenuButton>Newest</MenuButton>
           </div>
