+++ packages/app/pages/profile/index.tsx
@@ -1,7 +1,9 @@
 import Card from "@/components/Card";
+import Post, { PostList } from "@/components/Post";
+import TabBar, { TabButton } from "@/components/TabBar";
+import Title from "@/components/Title/Title";
 import UserCard from "@/components/UserCard";
 import { useWallet } from "@/components/WalletAuth";
-import { useEffect } from "react";
 
 const UnknownUser = () => {
   return (
@@ -12,7 +14,7 @@ const UnknownUser = () => {
 };
 
 const Profile = () => {
-  const { address } = useWallet();
+  const { address, ens } = useWallet();
 
   if (!address) {
     return <UnknownUser />;
@@ -21,8 +23,24 @@ const Profile = () => {
   return (
     <div>
       <Card>
-        <UserCard address={address} />
+        <Title>Profile</Title>
+        <UserCard address={address} ens={ens} />
       </Card>
+
+      <TabBar>
+        <TabButton active>Links</TabButton>
+        <div style={{ marginLeft: "auto" }}></div>
+      </TabBar>
+
+      <PostList>
+        <Post title="Post Title" />
+        <Post title="Design with Community in Mind: Cabin Core Contributor Mel Shields" />
+        <Post title="A brief history of decentralized cities and centralized states" />
+        <Post title="ConstitutionDAO: We Lost the Battle, But Will Win the War" />
+        <Post title="Growing the Writer’s Guild: Cabin Core Contributor Roxine Kee" />
+        <Post title="Building a Decentralized City: Cabin Core Contributor Phil Levin" />
+        <Post title="Around the Campfire, Cabin Contributor Jon Hillis" />
+      </PostList>
     </div>
   );
 };
